+++ dependency-reduced-pom.xml
@@ -28,7 +28,7 @@
                     <archive>
                         <manifest>
                             <addClasspath>true</addClasspath>
-                            <mainClass>ch.usi.si.codelounge.excel.ExcelParser</mainClass>
+                            <mainClass>ch.usi.si.codelounge.excel.Parser</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
